move math calculations to separate file
