+++ webapp/javascript/pages/math.spec.ts
@@ -1,7 +1,7 @@
-import { calculateMean, calculateStdDeviation } from './TagExplorerView';
+import { calculateMean, calculateStdDeviation } from './math';
 
-describe('TagExplorerView', () => {
-  describe('utils: calculateMean, calculateStdDeviation', () => {
+describe('math', () => {
+  describe('calculateMean, calculateStdDeviation', () => {
     test.each([
       [[1, 2, 3, 4, 5], 1.4142135623730951],
       [[23, 4, 6, 457, 65, 7, 45, 8], 145.13565852332775],
