+++ webapp/javascript/pages/math.ts
@@ -1,0 +1,15 @@
+export function calculateMean(arr: number[]) {
+  return arr.reduce((acc, b) => acc + b, 0) / arr.length;
+}
+
+export function calculateStdDeviation(array: number[], mean: number) {
+  const stdDeviation = Math.sqrt(
+    array.reduce((acc, b) => {
+      const dev = b - mean;
+
+      return acc + dev ** 2;
+    }, 0) / array.length
+  );
+
+  return stdDeviation;
+}
